+++ source/server/connection_handler_impl.cc
@@ -19,11 +19,11 @@ ConnectionHandlerImpl::ConnectionHandlerImpl(Event::Dispatcher& dispatcher,
     : dispatcher_(dispatcher), per_handler_stat_prefix_(per_handler_stat_prefix + "."),
       disable_listeners_(false) {}
 
-void ConnectionHandlerImpl::incNumConnections() { ++num_connections_; }
+void ConnectionHandlerImpl::incNumConnections() { ++num_handler_connections_; }
 
 void ConnectionHandlerImpl::decNumConnections() {
-  ASSERT(num_connections_ > 0);
-  --num_connections_;
+  ASSERT(num_handler_connections_ > 0);
+  --num_handler_connections_;
 }
 
 void ConnectionHandlerImpl::addListener(Network::ListenerConfig& config) {
@@ -81,12 +81,10 @@ void ConnectionHandlerImpl::enableListeners() {
   }
 }
 
-void ConnectionHandlerImpl::ActiveTcpListener::removeConnection(ActiveConnection& connection) {
+void ConnectionHandlerImpl::ActiveTcpListener::removeConnection(ActiveTcpConnection& connection) {
   ENVOY_CONN_LOG(debug, "adding to cleanup list", *connection.connection_);
-  ActiveConnectionPtr removed = connection.removeFromList(connections_);
+  ActiveTcpConnectionPtr removed = connection.removeFromList(connections_);
   parent_.dispatcher_.deferredDelete(std::move(removed));
-  ASSERT(parent_.num_connections_ > 0);
-  parent_.num_connections_--;
 }
 
 ConnectionHandlerImpl::ActiveListenerImplBase::ActiveListenerImplBase(
@@ -106,22 +104,24 @@ ConnectionHandlerImpl::ActiveListenerImplBase::ActiveListenerImplBase(
 ConnectionHandlerImpl::ActiveTcpListener::ActiveTcpListener(ConnectionHandlerImpl& parent,
                                                             Network::ListenerConfig& config)
     : ActiveTcpListener(
-          parent,
-          parent.dispatcher_.createListener(config.socket(), *this, config.bindToPort(),
-                                            config.handOffRestoredDestinationConnections()),
+          parent, parent.dispatcher_.createListener(config.socket(), *this, config.bindToPort()),
           config) {}
 
 ConnectionHandlerImpl::ActiveTcpListener::ActiveTcpListener(ConnectionHandlerImpl& parent,
                                                             Network::ListenerPtr&& listener,
                                                             Network::ListenerConfig& config)
     : ConnectionHandlerImpl::ActiveListenerImplBase(parent, std::move(listener), config),
-      parent_(parent) {}
+      parent_(parent) {
+  config.connectionBalancer().registerHandler(*this);
+}
 
 ConnectionHandlerImpl::ActiveTcpListener::~ActiveTcpListener() {
+  config_.connectionBalancer().unregisterHandler(*this);
+
   // Purge sockets that have not progressed to connections. This should only happen when
   // a listener filter stops iteration and never resumes.
   while (!sockets_.empty()) {
-    ActiveSocketPtr removed = sockets_.front()->removeFromList(sockets_);
+    ActiveTcpSocketPtr removed = sockets_.front()->removeFromList(sockets_);
     parent_.dispatcher_.deferredDelete(std::move(removed));
   }
 
@@ -130,12 +130,14 @@ ConnectionHandlerImpl::ActiveTcpListener::~ActiveTcpListener() {
   }
 
   parent_.dispatcher_.clearDeferredDeleteList();
-}
 
-Network::Listener*
-ConnectionHandlerImpl::findListenerByAddress(const Network::Address::Instance& address) {
-  Network::ConnectionHandler::ActiveListener* listener = findActiveListenerByAddress(address);
-  return listener ? listener->listener() : nullptr;
+  // By the time a listener is destroyed, in the common case, there should be no connections.
+  // However, this is not always true if there is an in flight rebalanced connection that is
+  // being posted. This assert is extremely useful for debugging the common path so we will leave it
+  // for now. If it becomes a problem (developers hitting this assert when using debug builds) we
+  // can revisit. This case, if it happens, should be benign on production builds. This case is
+  // covered in ConnectionHandlerTest::RemoveListenerDuringRebalance.
+  ASSERT(num_listener_connections_ == 0);
 }
 
 Network::ConnectionHandler::ActiveListener*
@@ -169,7 +171,7 @@ ConnectionHandlerImpl::findActiveListenerByAddress(const Network::Address::Insta
   return (listener_it != listeners_.end()) ? listener_it->second.get() : nullptr;
 }
 
-void ConnectionHandlerImpl::ActiveSocket::onTimeout() {
+void ConnectionHandlerImpl::ActiveTcpSocket::onTimeout() {
   listener_.stats_.downstream_pre_cx_timeout_.inc();
   ASSERT(inserted());
   ENVOY_LOG(debug, "listener filter times out after {} ms",
@@ -182,22 +184,22 @@ void ConnectionHandlerImpl::ActiveSocket::onTimeout() {
   unlink();
 }
 
-void ConnectionHandlerImpl::ActiveSocket::startTimer() {
+void ConnectionHandlerImpl::ActiveTcpSocket::startTimer() {
   if (listener_.listener_filters_timeout_.count() > 0) {
     timer_ = listener_.parent_.dispatcher_.createTimer([this]() -> void { onTimeout(); });
     timer_->enableTimer(listener_.listener_filters_timeout_);
   }
 }
 
-void ConnectionHandlerImpl::ActiveSocket::unlink() {
-  ActiveSocketPtr removed = removeFromList(listener_.sockets_);
+void ConnectionHandlerImpl::ActiveTcpSocket::unlink() {
+  ActiveTcpSocketPtr removed = removeFromList(listener_.sockets_);
   if (removed->timer_ != nullptr) {
     removed->timer_->disableTimer();
   }
   listener_.parent_.dispatcher_.deferredDelete(std::move(removed));
 }
 
-void ConnectionHandlerImpl::ActiveSocket::continueFilterChain(bool success) {
+void ConnectionHandlerImpl::ActiveTcpSocket::continueFilterChain(bool success) {
   if (success) {
     bool no_error = true;
     if (iter_ == accept_filters_.end()) {
@@ -230,13 +232,13 @@ void ConnectionHandlerImpl::ActiveSocket::continueFilterChain(bool success) {
     }
   }
 
-  // Filter execution concluded, unlink and delete this ActiveSocket if it was linked.
+  // Filter execution concluded, unlink and delete this ActiveTcpSocket if it was linked.
   if (inserted()) {
     unlink();
   }
 }
 
-void ConnectionHandlerImpl::ActiveSocket::newConnection() {
+void ConnectionHandlerImpl::ActiveTcpSocket::newConnection() {
   // Check if the socket may need to be redirected to another listener.
   ConnectionHandler::ActiveListener* new_listener = nullptr;
 
@@ -247,12 +249,17 @@ void ConnectionHandlerImpl::ActiveSocket::newConnection() {
   if (new_listener != nullptr) {
     // TODO(sumukhs): Try to avoid dynamic_cast by coming up with a better interface design
     ActiveTcpListener* tcp_listener = dynamic_cast<ActiveTcpListener*>(new_listener);
-    ASSERT(tcp_listener != nullptr, "ActiveSocket listener is expected to be tcp");
+    ASSERT(tcp_listener != nullptr, "ActiveTcpSocket listener is expected to be tcp");
     // Hands off connections redirected by iptables to the listener associated with the
     // original destination address. Pass 'hand_off_restored_destination_connections' as false to
-    // prevent further redirection.
-    tcp_listener->onAccept(std::move(socket_),
-                           false /* hand_off_restored_destination_connections */);
+    // prevent further redirection as well as 'rebalanced' as true since the connection has
+    // already been balanced if applicable inside onAcceptWorker() when the connection was
+    // initially accepted. Note also that we must account for the number of connections properly
+    // across both listeners.
+    // TODO(mattklein123): See note in ~ActiveTcpSocket() related to making this accounting better.
+    listener_.decNumConnections();
+    tcp_listener->incNumConnections();
+    tcp_listener->onAcceptWorker(std::move(socket_), false, true);
   } else {
     // Set default transport protocol if none of the listener filters did it.
     if (socket_->detectedTransportProtocol().empty()) {
@@ -264,10 +271,24 @@ void ConnectionHandlerImpl::ActiveSocket::newConnection() {
   }
 }
 
-void ConnectionHandlerImpl::ActiveTcpListener::onAccept(
-    Network::ConnectionSocketPtr&& socket, bool hand_off_restored_destination_connections) {
-  auto active_socket = std::make_unique<ActiveSocket>(*this, std::move(socket),
-                                                      hand_off_restored_destination_connections);
+void ConnectionHandlerImpl::ActiveTcpListener::onAccept(Network::ConnectionSocketPtr&& socket) {
+  onAcceptWorker(std::move(socket), config_.handOffRestoredDestinationConnections(), false);
+}
+
+void ConnectionHandlerImpl::ActiveTcpListener::onAcceptWorker(
+    Network::ConnectionSocketPtr&& socket, bool hand_off_restored_destination_connections,
+    bool rebalanced) {
+  if (!rebalanced) {
+    Network::BalancedConnectionHandler& target_handler =
+        config_.connectionBalancer().pickTargetHandler(*this);
+    if (&target_handler != this) {
+      target_handler.post(std::move(socket));
+      return;
+    }
+  }
+
+  auto active_socket = std::make_unique<ActiveTcpSocket>(*this, std::move(socket),
+                                                         hand_off_restored_destination_connections);
 
   // Create and run the filters
   config_.filterChainFactory().createListenerFilterChain(*active_socket);
@@ -293,54 +314,89 @@ void ConnectionHandlerImpl::ActiveTcpListener::newConnection(
   }
 
   auto transport_socket = filter_chain->transportSocketFactory().createTransportSocket(nullptr);
-  Network::ConnectionPtr new_connection =
-      parent_.dispatcher_.createServerConnection(std::move(socket), std::move(transport_socket));
-  new_connection->setBufferLimits(config_.perConnectionBufferLimitBytes());
+  ActiveTcpConnectionPtr active_connection(new ActiveTcpConnection(
+      *this,
+      parent_.dispatcher_.createServerConnection(std::move(socket), std::move(transport_socket)),
+      parent_.dispatcher_.timeSource()));
+  active_connection->connection_->setBufferLimits(config_.perConnectionBufferLimitBytes());
 
   const bool empty_filter_chain = !config_.filterChainFactory().createNetworkFilterChain(
-      *new_connection, filter_chain->networkFilterFactories());
+      *active_connection->connection_, filter_chain->networkFilterFactories());
   if (empty_filter_chain) {
-    ENVOY_CONN_LOG(debug, "closing connection: no filters", *new_connection);
-    new_connection->close(Network::ConnectionCloseType::NoFlush);
-    return;
+    ENVOY_CONN_LOG(debug, "closing connection: no filters", *active_connection->connection_);
+    active_connection->connection_->close(Network::ConnectionCloseType::NoFlush);
   }
-
-  onNewConnection(std::move(new_connection));
-}
-
-void ConnectionHandlerImpl::ActiveTcpListener::onNewConnection(
-    Network::ConnectionPtr&& new_connection) {
-  ENVOY_CONN_LOG(debug, "new connection", *new_connection);
 
   // If the connection is already closed, we can just let this connection immediately die.
-  if (new_connection->state() != Network::Connection::State::Closed) {
-    ActiveConnectionPtr active_connection(
-        new ActiveConnection(*this, std::move(new_connection), parent_.dispatcher_.timeSource()));
+  if (active_connection->connection_->state() != Network::Connection::State::Closed) {
+    ENVOY_CONN_LOG(debug, "new connection", *active_connection->connection_);
+    active_connection->connection_->addConnectionCallbacks(*active_connection);
     active_connection->moveIntoList(std::move(active_connection), connections_);
-    parent_.num_connections_++;
   }
 }
 
-ConnectionHandlerImpl::ActiveConnection::ActiveConnection(ActiveTcpListener& listener,
-                                                          Network::ConnectionPtr&& new_connection,
-                                                          TimeSource& time_source)
+namespace {
+// Structure used to allow a unique_ptr to be captured in a posted lambda. See below.
+struct RebalancedSocket {
+  Network::ConnectionSocketPtr socket;
+};
+using RebalancedSocketSharedPtr = std::shared_ptr<RebalancedSocket>;
+} // namespace
+
+void ConnectionHandlerImpl::ActiveTcpListener::post(Network::ConnectionSocketPtr&& socket) {
+  // It is not possible to capture a unique_ptr because the post() API copies the lambda, so we must
+  // bundle the socket inside a shared_ptr that can be captured.
+  // TODO(mattklein123): It may be possible to change the post() API such that the lambda is only
+  // moved, but this is non-trivial and needs investigation.
+  RebalancedSocketSharedPtr socket_to_rebalance = std::make_shared<RebalancedSocket>();
+  socket_to_rebalance->socket = std::move(socket);
+
+  parent_.dispatcher_.post([socket_to_rebalance, tag = config_.listenerTag(), &parent = parent_]() {
+    // TODO(mattklein123): We should probably use a hash table here to lookup the tag instead of
+    // iterating through the listener list.
+    for (const auto& listener : parent.listeners_) {
+      if (listener.second->listener() != nullptr && listener.second->listenerTag() == tag) {
+        // TODO(mattklein123): Remove dynamic_cast here in follow up cleanup.
+        ActiveTcpListener& tcp_listener = *dynamic_cast<ActiveTcpListener*>(listener.second.get());
+        tcp_listener.onAcceptWorker(std::move(socket_to_rebalance->socket),
+                                    tcp_listener.config_.handOffRestoredDestinationConnections(),
+                                    true);
+        return;
+      }
+    }
+  });
+}
+
+ConnectionHandlerImpl::ActiveTcpConnection::ActiveTcpConnection(
+    ActiveTcpListener& listener, Network::ConnectionPtr&& new_connection, TimeSource& time_source)
     : listener_(listener), connection_(std::move(new_connection)),
       conn_length_(new Stats::Timespan(listener_.stats_.downstream_cx_length_ms_, time_source)) {
   // We just universally set no delay on connections. Theoretically we might at some point want
   // to make this configurable.
   connection_->noDelay(true);
-  connection_->addConnectionCallbacks(*this);
+
   listener_.stats_.downstream_cx_total_.inc();
   listener_.stats_.downstream_cx_active_.inc();
   listener_.per_worker_stats_.downstream_cx_total_.inc();
   listener_.per_worker_stats_.downstream_cx_active_.inc();
+
+  // Active connections on the handler (not listener). The per listener connections have already
+  // been incremented at this point either via the connection balancer or in the socket accept
+  // path if there is no configured balancer.
+  ++listener_.parent_.num_handler_connections_;
 }
 
-ConnectionHandlerImpl::ActiveConnection::~ActiveConnection() {
+ConnectionHandlerImpl::ActiveTcpConnection::~ActiveTcpConnection() {
   listener_.stats_.downstream_cx_active_.dec();
   listener_.stats_.downstream_cx_destroy_.inc();
   listener_.per_worker_stats_.downstream_cx_active_.dec();
   conn_length_->complete();
+
+  // Active listener connections (not handler).
+  listener_.decNumConnections();
+
+  // Active handler connections (not listener).
+  listener_.parent_.decNumConnections();
 }
 
 ActiveUdpListener::ActiveUdpListener(Network::ConnectionHandler& parent,